rename and fix paths
