+++ test/pop.spec.js
@@ -1,4 +1,4 @@
-import {default as pop} from '../src/immutable-pop';
+import pop from '../src/pop';
 import chai from 'chai';
 
 const expect = chai.expect;
