+++ test/push.spec.js
@@ -1,4 +1,4 @@
-import {default as push} from '../src/immutable-push';
+import push from '../src/push';
 import chai from 'chai';
 
 const expect = chai.expect;
